Remove SSR, inline CSS & JS in production
