+++ client/containers/Layout.js
@@ -11,7 +11,7 @@ import Header from '../components/Header'
 import Footer from '../components/Footer'
 import Copyright from '../components/Copyright'
 
-if ( process.browser && 'scrollRestoration' in window.history) {
+if ( 'scrollRestoration' in window.history) {
   window.history.scrollRestoration = 'manual';
 }
