+++ client/main.js
@@ -1,29 +1,15 @@
 import React from 'react'
 import { render } from 'react-dom'
-import { StyleSheet } from 'aphrodite/no-important'
 import nprogress from 'nprogress'
 
 import configureStore from './store/configureStore'
 import App from './containers/App'
 import './services/ga'
-import preserver from './routes/Preserver'
 
 // Hide spinner from nprogress
 nprogress.configure({
   showSpinner: false
 });
-
-// Re-hydrate Aphrodite from server-generated class names
-const rehydrateFrom = document.getElementById('aphro-hydrate');
-if ( rehydrateFrom ) {
-  StyleSheet.rehydrate(JSON.parse(rehydrateFrom.innerHTML));
-}
-
-// Re-hydrate routes HTML
-const routesRootEl = document.getElementById('lwjgl-routes');
-if ( routesRootEl && process.env.NODE_ENV === 'production' ) {
-  preserver.store(routesRootEl.innerHTML);
-}
 
 // Configure Redux store
 const store = configureStore();
