+++ client/services/supports-passive.js
@@ -1,16 +1,13 @@
 let supportsPassive = false;
 
-if ( process.browser ) {
-  try {
-    let opts = Object.defineProperty({}, 'passive', {
-      get: function() {
-        supportsPassive = true;
-      }
-    });
-    //noinspection JSCheckFunctionSignatures
-    window.addEventListener('test', null, opts);
-  } catch (ignore) {
-  }
-}
+try {
+  let opts = Object.defineProperty({}, 'passive', {
+    get: function() {
+      supportsPassive = true;
+    }
+  });
+  //noinspection JSCheckFunctionSignatures
+  window.addEventListener('test', null, opts);
+} catch (ignore) {}
 
 export default supportsPassive;
