+++ package.json
@@ -69,7 +69,6 @@
     "babel-runtime": "6.18.0",
     "bootstrap": "4.0.0-alpha.5",
     "chalk": "1.1.3",
-    "chunk-manifest-webpack-plugin": "github:apostolos/chunk-manifest-webpack-plugin",
     "classnames": "2.2.5",
     "cli-table": "0.3.1",
     "cross-env": "3.1.3",
