+++ server/views/layout.pug
@@ -5,15 +5,11 @@ html(lang="en-US")
 
   head
     meta(charset="utf-8")
-    if ( head && head.title )
-      !=head.title
-    else
-      title LWJGL - Lightweight Java Game Library
-    link(rel="preload" as="style" href=`/css/${css}`)
-    link(rel="preload" as="script" href=`/js/${bundle}`)
+    title LWJGL - Lightweight Java Game Library
+
     if ( chunk )
       link(rel="preload" as="script" href=`/js/${chunk}`)
-    link(rel="stylesheet" href=`/css/${css}`)
+
     meta(name="viewport" content="width=device-width,initial-scale=1,user-scalable=no")
     meta(name="theme-color" content="#1b2426")
     meta(name="application-name" content="LWJGL")
@@ -30,25 +26,10 @@ html(lang="en-US")
     link(rel="icon" type="image/png" href="/img/fav/favicon-32x32.png" sizes="32x32")
     link(rel="manifest" href="/manifest.json")
 
-    if ( head )
-      !=head.meta
-      !=head.link
-
-    if ( aphrodite && aphrodite.content.length )
-      style(data-aphrodite="")!=aphrodite.content
+    if ( inline )
+      style !{css}
+    else
+      link(rel="stylesheet" href=`/css/${css}`)
 
   body(class=bodyClass)
-
     block content
-
-    if ( renderApp )
-      if ( aphrodite )
-        script(id="aphro-hydrate" type="application/json")!=JSON.stringify(aphrodite.renderedClassNames)
-      if ( chunks )
-        script!=`manifest=${chunks};`
-      if ( ie )
-        //- https://polyfill.io/v2/docs/features/
-        script(src="https://cdn.polyfill.io/v2/polyfill.min.js?flags=gated&features=requestAnimationFrame,Element.prototype.classList")
-        script.
-          for ( var i in manifest ) if ( manifest.hasOwnProperty(i) ) manifest[i] += '?v=' + new Date().getTime();
-      script(src=`/js/${bundle}` async defer)
